Set up webserver and start/setup scripts
This commit creates a static express server to host the built version of the
React frontend. It sets up a project-wide setup and start script at the root
directory, and keeps the development start in the subdirs.

The server's development port is 3000 while the React development server uses
port 3001 as they are expected to coexist while working locally. In production
only the server exists as it's hosting an optimized, static version of the
front end (technically available in local version too but not useful).